requirements.txt
streamlit
langchain-community
langchain-chroma
langchain_ollama
tempfile
ollama
pypdf
